pyproject.toml
[tool.poetry]
name = "flask-lesson"
version = "0.1.0"
description = ""
authors = ["Nik Bezmaternykh <nbezmaternykh@7tech.ru>"]
readme = "README.md"
packages = [{include = "flask_lesson"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
sqlalchemy = "^2.0.23"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
